functionalize rssfeed return and integrate
